#9. 1) 도메인에 맞게 위치 변경     2) 좋아요 생성 조건 체크     3) 북마크 계층 관계 수정
