+++ src/main/java/com/zerozone/vintage/account/Account.java
@@ -1,5 +1,7 @@
-package com.zerozone.vintage.domain;
+package com.zerozone.vintage.account;
 
+import com.zerozone.vintage.tag.CameraTag;
+import com.zerozone.vintage.tag.LocationTag;
 import jakarta.persistence.*;
 import lombok.*;
