+++ src/main/java/com/zerozone/vintage/account/UserAccount.java
@@ -1,6 +1,5 @@
 package com.zerozone.vintage.account;
 
-import com.zerozone.vintage.domain.Account;
 import lombok.Getter;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
