+++ src/main/java/com/zerozone/vintage/board/BoardRepository.java
@@ -1,6 +1,5 @@
 package com.zerozone.vintage.board;
 
-import com.zerozone.vintage.domain.Board;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
