+++ src/main/java/com/zerozone/vintage/board/BoardService.java
@@ -1,7 +1,6 @@
 package com.zerozone.vintage.board;
 
-import com.zerozone.vintage.domain.Account;
-import com.zerozone.vintage.domain.Board;
+import com.zerozone.vintage.account.Account;
 import com.zerozone.vintage.exception.CustomException;
 import java.time.LocalDateTime;
 import java.util.List;
