+++ src/main/java/com/zerozone/vintage/bookmark/Bookmark.java
@@ -1,10 +1,11 @@
-package com.zerozone.vintage.domain;
+package com.zerozone.vintage.bookmark;
 
+import com.zerozone.vintage.account.Account;
+import com.zerozone.vintage.board.Board;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
-import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -17,7 +18,7 @@ import lombok.Setter;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class Comment {
+public class Bookmark {
 
     @Id
     @GeneratedValue
@@ -27,11 +28,5 @@ public class Comment {
     private Board board;
 
     @ManyToOne
-    private Account author;
-
-    private String content;
-
-    private LocalDateTime createdDateTime;
-
-    private LocalDateTime updatedDateTime;
+    private Account account;
 }
