+++ src/main/java/com/zerozone/vintage/bookmark/BookmarkRepository.java
@@ -1,6 +1,5 @@
 package com.zerozone.vintage.bookmark;
 
-import com.zerozone.vintage.domain.Bookmark;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
