+++ src/main/java/com/zerozone/vintage/comment/Comment.java
@@ -1,9 +1,12 @@
-package com.zerozone.vintage.domain;
+package com.zerozone.vintage.comment;
 
+import com.zerozone.vintage.account.Account;
+import com.zerozone.vintage.board.Board;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
+import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -16,7 +19,7 @@ import lombok.Setter;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class Bookmark {
+public class Comment {
 
     @Id
     @GeneratedValue
@@ -26,5 +29,11 @@ public class Bookmark {
     private Board board;
 
     @ManyToOne
-    private Account account;
+    private Account author;
+
+    private String content;
+
+    private LocalDateTime createdDateTime;
+
+    private LocalDateTime updatedDateTime;
 }
