+++ src/main/java/com/zerozone/vintage/comment/CommentRepository.java
@@ -1,6 +1,5 @@
 package com.zerozone.vintage.comment;
 
-import com.zerozone.vintage.domain.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
