+++ src/main/java/com/zerozone/vintage/tag/CameraTag.java
@@ -1,4 +1,4 @@
-package com.zerozone.vintage.domain;
+package com.zerozone.vintage.tag;
 
 
 import jakarta.persistence.Column;
@@ -14,7 +14,7 @@ import lombok.*;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class LocationTag {
+public class CameraTag {
 
     @Id
     @GeneratedValue
