+++ src/main/java/com/zerozone/vintage/tag/CameraTagRepository.java
@@ -1,6 +1,5 @@
 package com.zerozone.vintage.tag;
 
-import com.zerozone.vintage.domain.CameraTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
