+++ src/main/java/com/zerozone/vintage/tag/LocationTagRepository.java
@@ -1,7 +1,5 @@
 package com.zerozone.vintage.tag;
 
-import com.zerozone.vintage.domain.CameraTag;
-import com.zerozone.vintage.domain.LocationTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
