+++ src/main/java/com/zerozone/vintage/user/UserApiController.java
@@ -3,9 +3,9 @@ package com.zerozone.vintage.user;
 import com.zerozone.vintage.account.AccountService;
 import com.zerozone.vintage.account.AuthenticationManager;
 import com.zerozone.vintage.account.CheckedUser;
-import com.zerozone.vintage.domain.Account;
-import com.zerozone.vintage.domain.CameraTag;
-import com.zerozone.vintage.domain.LocationTag;
+import com.zerozone.vintage.account.Account;
+import com.zerozone.vintage.tag.CameraTag;
+import com.zerozone.vintage.tag.LocationTag;
 import com.zerozone.vintage.dto.CustomResDto;
 import com.zerozone.vintage.exception.CustomException;
 import com.zerozone.vintage.tag.CameraTagRepository;
