+++ src/main/java/com/zerozone/vintage/user/UserController.java
@@ -3,7 +3,7 @@ package com.zerozone.vintage.user;
 import com.zerozone.vintage.account.AccountRepository;
 import com.zerozone.vintage.account.AccountService;
 import com.zerozone.vintage.account.CheckedUser;
-import com.zerozone.vintage.domain.Account;
+import com.zerozone.vintage.account.Account;
 import com.zerozone.vintage.exception.CustomException;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
