+++ src/test/java/com/zerozone/vintage/like/LikeDislikeApiControllerTest.java
@@ -9,9 +9,8 @@ import com.zerozone.vintage.account.AccountRepository;
 import com.zerozone.vintage.account.AccountService;
 import com.zerozone.vintage.account.SignUpForm;
 import com.zerozone.vintage.board.BoardRepository;
-import com.zerozone.vintage.domain.Account;
-import com.zerozone.vintage.domain.Board;
-import com.zerozone.vintage.domain.LikeDislike;
+import com.zerozone.vintage.account.Account;
+import com.zerozone.vintage.board.Board;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
