+++ src/test/java/com/zerozone/vintage/user/UserControllerTest.java
@@ -3,9 +3,9 @@ package com.zerozone.vintage.user;
 import com.zerozone.vintage.account.AccountRepository;
 import com.zerozone.vintage.account.AccountService;
 import com.zerozone.vintage.account.SignUpForm;
-import com.zerozone.vintage.domain.Account;
-import com.zerozone.vintage.domain.CameraTag;
-import com.zerozone.vintage.domain.LocationTag;
+import com.zerozone.vintage.account.Account;
+import com.zerozone.vintage.tag.CameraTag;
+import com.zerozone.vintage.tag.LocationTag;
 import com.zerozone.vintage.tag.CameraTagRepository;
 import com.zerozone.vintage.tag.LocationTagRepository;
 import org.junit.jupiter.api.BeforeEach;
